requirements.txt
djangorestframework==3.12.1
requests
pillow
